+++ apps/prairielearn/src/components/TagsTopicsTable.tsx
@@ -1,4 +1,4 @@
-import { useState } from 'preact/compat';
+import { useState } from 'react';
 
 import { type StaffTag, type StaffTopic } from '../lib/client/safe-db-types.js';
 import { ColorJsonSchema } from '../schemas/infoCourse.js';
@@ -126,7 +126,7 @@ export function TagsTopicsTable<Entity extends StaffTag | StaffTopic>({
             <thead>
               <tr>
                 {editMode && allowEdit && (
-                  <th style="width: 1%">
+                  <th style={{ width: '1%' }}>
                     <span className="visually-hidden">Edit and Delete</span>
                   </th>
                 )}